+++ src/test/kotlin/coffeelisp/MathTest.kt
@@ -8,10 +8,10 @@ import kotlin.test.assertEquals
 class MathTest {
     @Test
     fun testSubtraction() {
-        assertEquals(LispNumber(0), interpret("(sub1 1)"))
+        assertEquals(LispNumber(0.toBigInteger()), interpret("(sub1 1)"))
     }
 
     @Test fun testAddition() {
-        assertEquals(LispNumber(1), interpret("(add1 0)"))
+        assertEquals(LispNumber(1.toBigInteger()), interpret("(add1 0)"))
     }
 }
